+++ src/components/NavBar/SettingsModal/index.jsx
@@ -1,8 +1,10 @@
 import { useContext, useEffect, useState } from "react"
 import { GameContext } from "../../../context/GameContext"
+import { DataContext } from "../../../context/DataContext"
 
 const SettingsModal = () => {
-    const {toggleTheme, theme, bgModal, bgColor, resetGame} = useContext(GameContext)
+    const {theme, bgModal, bgColor} = useContext(DataContext)
+    const {resetGame} = useContext(GameContext)
 
 
     return (
@@ -10,6 +12,7 @@ const SettingsModal = () => {
             Settings
             <div className="flex flex-row">
                 <button className={`w-16 h-8 rounded-xl flex items-center justify-center ${theme == 'light' ? 'hover:bg-zinc-300' : 'hover:bg-zinc-700'} ` + bgModal} onClick={() => resetGame()}>Reset</button>
+                <button className={`w-16 h-8 rounded-xl flex items-center justify-center ${theme == 'light' ? 'hover:bg-zinc-300' : 'hover:bg-zinc-700'} ` + bgModal} onClick={() => saveGame()}>Import</button>
             </div>
         </div>
     )