+++ CHANGELOG.md
@@ -7,13 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+## [1.0.1] - 2023-08-16
+
+### Changed
+- Split the gameContext in two files, game context for the logical context and data context to data.
+- glass cannon changed
+
+### fixed
+- Necromancer bug
+
 
 ### Changed
 - Split the gameContext in two files, game context for the logical context and data context to data.
 - Changed the time function
 
-
 ## [1.0.0] - 2023-08-14
 - First stable version
+
+
+
+